refactor(benchmark): restructure benchmark layout
