+++ dev-docs/CONTRIBUTING.md
@@ -359,7 +359,7 @@ The code generator lives in the [`src/generator`](../src/generator) sub-folder w
 The implementation that we have right now is being refactored to produce FunC ASTs and then pretty-print those ASTs as strings instead of producing source FunC code in one step. Here is the relevant pull request: <https://github.com/tact-lang/tact/pull/559>.
 
 One can find the end-to-end codegen test spec files in the [`src/test/e2e-emulated`](../src/test/e2e-emulated) folder. The test contracts are located in the subfolders of the [`src/test/e2e-emulated`](../src/test/e2e-emulated) folder. Many of those spec files test various language features in relative isolation.
-An important spec file that tests argument passing semantics for functions and assignment semantics for variables is here: [`src/test/e2e-emulated/semantics.spec.ts`](../src/test/e2e-emulated/functions/semantics.spec.ts).
+An important spec file that tests argument passing semantics for functions and assignment semantics for variables is here: [`src/test/e2e-emulated/functions/semantics.spec.ts`](../src/test/e2e-emulated/functions/semantics.spec.ts).
 
 Contracts with `inline` in the name of the file set `experimental.inline` config option to `true`.
 Contracts with `external` in the name of the file set the `external` config option to `true`.
@@ -384,9 +384,10 @@ Benchmarks are located inside `src/benchmarks/`, one directory per benchmark:
 | ------------------------ | -------------------------------------------------------- |
 | `tact/`                  | Tact project that is being benchmarked                   |
 | `func/`                  | Equivalent FunC project that we compare against          |
-| `<benchmark>.spec.ts`    | Jest test spec that prepares and runs the benchmark      |
-| `results_gas.json`       | Aggregated gas‑consumption results, updated by the CLI   |
-| `results_code_size.json` | Contract byte‑code size history, also updated by the CLI |
+| `test.spec.ts`           | Jest test spec for contract functionality testing        |
+| `bench.spec.ts`          | Jest test spec for performance benchmarking              |
+| `gas.json`               | Aggregated gas‑consumption results, updated by the CLI   |
+| `size.json`              | Contract byte‑code size history, also updated by the CLI |
 
 > **CLI support** – All commands for creating, updating, or comparing benchmarks are documented in the [Updating Benchmarks](#benchmarks) section.
