+++ src/benchmarks/update.build.ts
@@ -216,7 +216,7 @@ const updateCodeSizeResultsFile = async (
 
 const main = async () => {
     try {
-        const benchmarkPaths = globSync(["**/*.spec.ts"], {
+        const benchmarkPaths = globSync(["**/bench.spec.ts"], {
             cwd: __dirname,
         });
 
@@ -237,11 +237,11 @@ const main = async () => {
                 return;
             }
 
-            const resultsGas = join(specPath, "..", "results_gas.json");
+            const resultsGas = join(specPath, "..", "gas.json");
             const resultsCodeSize = join(
                 specPath,
                 "..",
-                "results_code_size.json",
+                "size.json",
             );
 
             const isUpdate = typeof process.env.ADD !== "undefined";
